+++ labbook_FE/default_settings.py
@@ -1,3 +1,3 @@
 # -*- coding:utf-8 -*-
 
-APP_VERSION = '3.3.1'
+APP_VERSION = '3.3.2'